модуляция файлов
